Move mocks to __test__ dir
